# Conflicts:
#	app/Http/Controllers/NotesController.php
#	app/Http/Controllers/UserController.php
#	app/Listeners/StoreUserLoginHistory.php
#	app/Livewire/Administrators/GroupManagement.php
#	app/Livewire/Administrators/LibrarianManagement.php
#	app/Livewire/Administrators/StudentManagement.php
#	app/Livewire/Administrators/TeacherManagement.php
#	app/Livewire/Administrators/UserLoginLog.php
#	app/Livewire/Notes/ShareNoteRecipientsSelect.php
#	app/Livewire/Students/DocumentGenerator.php
#	app/Livewire/Students/Overview.php
#	app/Livewire/Teachers/StudentLogin.php
#	app/Models/LoginActivity.php
#	app/Models/Teacher.php
#	app/Providers/EventServiceProvider.php
#	app/Services/UserLoginService.php
#	app/Traits/BelongsToSchoolEnhanced.php
#	app/Traits/ShouldScopeSchool.php
#	resources/js/bootstrap.js
#	resources/views/components/layouts/app.blade.php
#	resources/views/livewire/administrators/school-switcher.blade.php
#	resources/views/notes/show.blade.php
#	routes/administrator.php
#	routes/web.php

+++ database/migrations/2025_07_17_005644_create_parent_student_table.php
@@ -20,9 +20,15 @@ return new class extends Migration
         });
 
         Schema::table('parents', static function (Blueprint $table) {
-            $table->dropForeign(['student_id']);
-            $table->dropColumn('student_id');
-            $table->dropColumn('relationship');
+            if (Schema::hasColumn('parents', 'student_id')) {
+                $table->dropIndex(['student_id']);
+                $table->dropForeign(['student_id']);
+                $table->dropColumn('student_id');
+            }
+            if (Schema::hasColumn('parents', 'relationship')) {
+                $table->dropIndex(['relationship']);
+                $table->dropColumn('relationship');
+            }
         });
     }
